pyqt-pdf-analyzer/requirements.txt
PyQt6>=6.4.0
PyMuPDF>=1.23.0
pandas>=1.5.0
Pillow>=9.0.0
